Support for controllerAs syntax
The original code assumes that ng-model does not contain a ‘.’ or does not use controllerAs syntax. This means using ng-model like vm.foo or $ctrl.foo will not work as it will search for scope[vm.foo] instead of scope.vm.foo. Using a path function inspired from lodash, dynamically retrieve value on scope with the given ng-model as the path.